+++ playwright/pageobjects/congrats.page.ts
@@ -1,4 +1,5 @@
 import { Page, Locator } from "@playwright/test";
+import { TEST_PATRON_INFO, TEST_BARCODE_NUMBER } from "../utils/constants";
 
 export class CongratsPage {
   readonly page: Page;
@@ -6,6 +7,7 @@ export class CongratsPage {
   readonly stepHeading: Locator;
   readonly getStartedHeading: Locator;
   readonly memberNameHeading: Locator;
+  readonly memberName: Locator;
   readonly issuedDateHeading: Locator;
   readonly issuedDate: Locator;
   readonly barcodeNumber: Locator;
@@ -34,8 +36,15 @@ export class CongratsPage {
     this.memberNameHeading = page
       .locator("#member-name")
       .getByText("MEMBER NAME");
+    this.memberName = page.getByText(
+      `${TEST_PATRON_INFO.firstName} ${TEST_PATRON_INFO.lastName}`,
+      { exact: true }
+    );
     this.issuedDateHeading = page.locator("#issued").getByText("ISSUED");
     this.issuedDate = page.locator("#issued").getByText(this.getDate(), {
+      exact: true,
+    });
+    this.barcodeNumber = page.getByText(TEST_BARCODE_NUMBER, {
       exact: true,
     });
     this.libraryCardBackground = page.locator(".background-lion");
@@ -62,15 +71,9 @@ export class CongratsPage {
 
   public getDate(): string {
     const today = new Date();
-
-    // NOTE: .toString() is sufficient as it doesn't add leading zeros.
-    const month = today.getMonth() + 1; // Month is 1-indexed
-    const day = today.getDate(); // Day is 1-31
-    const year = today.getFullYear();
-
-    // Concatenates without padding, matching the 11/4/2025 format
-    const formattedDate = `${month}/${day}/${year}`;
-
-    return formattedDate; // Example output: 11/4/2025
+    const formattedDate = `${(today.getMonth() + 1).toString()}/${today
+      .getDate()
+      .toString()}/${today.getFullYear()}`;
+    return formattedDate; // M/D/YYYY
   }
 }